Restructure create-keystone-app package
