+++ .changeset/09f7835c/changes.md
@@ -1,0 +1,1 @@
+- Refactor internal code for future testability
